Refactored share modal files
